pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "castleclashclient"
dynamic = ["version"]
description = 'Unofficial Castle Clash client written in Python 3'
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
keywords = []
authors = [
  { name = "anthy", email = "126660548+yntha@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pycryptodome",
  "colorama",
  "aiohttp",
  "xmltodict"
]

[project.urls]
Documentation = "https://github.com/yntha/castleclashclient#readme"
Issues = "https://github.com/yntha/castleclashclient/issues"
Source = "https://github.com/yntha/castleclashclient"

[project.scripts]
castleclashclient = "castleclashclient.__main__:main"

[tool.hatch.version]
path = "src/castleclashclient/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/castleclashclient tests}"
style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py311"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003", "DTZ007", "T201",
  # Ignore checks for possible passwords
  "S105", "S106", "S107", "PLW0127", "S101", "N801", "N817", "S314", "PLR2004", "FBT001", "FBT002", "S311",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "PLR5501", "B904", "EM102", "B905", "EM101"
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["castleclashclient"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["castleclashclient", "tests"]
branch = true
parallel = true
omit = [
  "src/castleclashclient/__about__.py",
]

[tool.coverage.paths]
castleclashclient = ["src/castleclashclient", "*/castleclashclient/src/castleclashclient"]
tests = ["tests", "*/castleclashclient/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
